Changes
There was a bug in the TimerService. Somehow, the TimerService module occasionally tried to schedule timers
to the past. Consequently, the omnet++ core raised an exception, due to the whole simulation failed. My
assumption is that the root cause is some kind of precision or random number generation issue. The fix checks
if the timer to be scheduled is valid. If not, it schedules the timer to SimTime().

 Changes to be committed:
	modified:   TimerService.cc